requirements.txt
discord.py
mysql-connector-python
python-dotenv
pytz